pi expansion for bigger cases fixed

+++ IrrationalNumbers/IrrationalNumbers.Logic/Expansions/HyperbolicTangentExpansion.cs
@@ -10,7 +10,7 @@ namespace IrrationalNumbers.Logic.Expansions
             _exponentExpansion = new ExponentTaylorExpansion();
         }
 
-        public BigDecimal ExpandFunction(int wantedRemainder, double x)
+        public BigDecimal ExpandFunction(int wantedRemainder, BigDecimal x)
         {
             var ex = _exponentExpansion.ExpandFunction(wantedRemainder, x);
             var minusEx = _exponentExpansion.ExpandFunction(wantedRemainder, -x);
